Switched around the bit values in gain
They seemed a bit arbitrary. Now the more gain you have the more bits it gives which makes sense in my head seeing as this is an amplifier. The good news is that this resulted in a weight in grams that's within the tolerance of my scale (50 kg scale measuring an item that weighs 1876g at values between 1700 and 2000 is good enough since it's around1.5% and this is a cheap part from China)